結算頁統計、修正輸入錯誤加2coin之bug

+++ Assets/Scripts/Character.cs
@@ -61,7 +61,7 @@ public class Character : MonoBehaviour {
 			}
 			else if(transform.name == "Player") {
 				PlayerPrefs.SetFloat("FinalATK", atk);
-				PlayerPrefs.SetInt("FinalSPD", spd);
+				PlayerPrefs.SetFloat("FinalSPD", spd);
 			}
 			Destroy (gameObject);
 		}
